Move net_http integration tests to the appropriate file
